Format strings checked by gcc & clang, crash fix
* TestVar segfaulted on ubuntu-latest and passed on Rocky 9.
  The units list refusal in _intr_unitsOk had a %s and no
  argument, so vsnprintf read a stale pointer

* das_error_func and daslog now carry the GNU printf format
  attribute via _das_fmt_check() in defs.h

* das_error_str() macro added for fixed strings with no
  formatting.

* das_cred_init on longer leaks information on a NULL argument

Developed-with: Claude Fable 5.1

+++ das3/stream.c
@@ -713,8 +713,8 @@ DasStream* new_DasStream_str(DasBuf* pBuf, int nModel)
 
 	if(!nParRet){
 		/* Report before the free: the line number and error code live in the parser */
-		das_error(DASERR_STREAM, "Parse error at line %d:\n%s\n",
-		           XML_GetCurrentLineNumber(p), XML_ErrorString(XML_GetErrorCode(p))
+		das_error(DASERR_STREAM, "Parse error at line %lu:\n%s\n",
+		           (unsigned long)XML_GetCurrentLineNumber(p), XML_ErrorString(XML_GetErrorCode(p))
 		);
 		XML_ParserFree(p);
 		del_DasStream(pThis);           // Don't leak on fail
